requirements.txt
amqp==2.6.1
asgiref==3.2.10
autopep8==1.5.4
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
Django==3.0.3
django-cors-headers==3.4.0
djangorestframework==3.11.0
drf-writable-nested==0.6.0
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
unicodecsv==0.14.1
urllib3==1.25.10
vine==1.3.0
yapf==0.30.0
zipp==3.1.0
